Add IP GeoLocation
git-svn-id: file:///home/svn/DVA2006@5920 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/commands/register/RegisterCommand.java
@@ -26,7 +26,7 @@ import org.deltava.util.system.SystemData;
 /**
  * A Web Site Command to register a new Applicant.
  * @author Luke
- * @version 2.4
+ * @version 2.5
  * @since 1.0
  */
 
@@ -217,7 +217,7 @@ public class RegisterCommand extends AbstractCommand {
 			Collection<RegistrationBlock> regBList = sysdao.getBlocks();
 			
 			// Check the blacklist
-			int regAddr = NetworkUtils.pack(a.getRegisterAddress());
+			long regAddr = NetworkUtils.pack(a.getRegisterAddress());
 			for (Iterator<RegistrationBlock> i = regBList.iterator(); i.hasNext(); ) {
 				RegistrationBlock rb = i.next();
 				boolean doBlock = false;
